IRK Refresh issue: IRK refreshing even device is in config file
Issue:
In General if there is any bonded device IRK shouldnt refresh.
But IRK is refreshing even at least one bonded device is in
the configuration file.

Fix:
We are not reading the LE keys to get num of bonded devices.
It will just calculate the number of sections and returns
from it.

CRs-Fixed: 3200398

Change-Id: I97f656edfc38355bfea5f0a570549675d33744bb