GATT: Handle Gatt Read multiple response for extended properties
Precondition:
Remote with Gatt Extended Property Descriptor is required.

Use case :
1.Keep peripheral device in advertising mode.
2.Pair and connect peripheral with central device.

Issue:
Remote services are not updated to applications

Root Cause:
Gatt database is not getting built if remote is
responding with error for Read multiple request of
GATT Ext Prop descriptor

Fix:
When remote is responding with error or wrong size data for
Gatt Extended Property Descriptor, skip reading remaining
handles for it and build the GATT database.

CRs-Fixed: 3213831

Change-Id: I29cece9c719bec420a4ad06343031de7cf771f41