Add an unordered set to track the alive ClientInterface objects
This change will import an unordered set to track all the alive
BluetoothAudioClientInterface objects. It will insert the object
to the set in the constructor and remove the object from set in
the destructor. In the binder death callback, it need to check
whether the object is still alive before restarting connection
with new Audio Hal.

Change-Id: I0675337da6c9e66f5f67db8c75b2ac97c414ef7a
CRs-fixed: 3240549